+++ json_serializable/example/example.dart
@@ -6,13 +6,10 @@ import 'package:json_annotation/json_annotation.dart';
 
 part 'example.g.dart';
 
-@JsonSerializable()
+@JsonSerializable(createJsonSchema: true)
 class Person {
-  /// The generated code assumes these values exist in JSON.
   final String firstName, lastName;
 
-  /// The generated code below handles if the corresponding JSON value doesn't
-  /// exist or is empty.
   final DateTime? dateOfBirth;
 
   Person({required this.firstName, required this.lastName, this.dateOfBirth});
@@ -23,4 +20,7 @@ class Person {
 
   /// Connect the generated [_$PersonToJson] function to the `toJson` method.
   Map<String, dynamic> toJson() => _$PersonToJson(this);
+
+  /// The JSON Schema for this class.
+  static const jsonSchema = _$PersonJsonSchema;
 }