+++ json_serializable/lib/src/shared_checkers.dart
@@ -13,6 +13,12 @@ const coreStringTypeChecker = TypeChecker.fromUrl('dart:core#String');
 
 const coreMapTypeChecker = TypeChecker.fromUrl('dart:core#Map');
 
+const coreDateTimeTypeChecker = TypeChecker.fromUrl('dart:core#DateTime');
+
+const coreUriTypeChecker = TypeChecker.fromUrl('dart:core#Uri');
+
+const coreBigIntTypeChecker = TypeChecker.fromUrl('dart:core#BigInt');
+
 /// Returns the generic type of the [Iterable] represented by [type].
 ///
 /// If [type] does not extend [Iterable], an error is thrown.