+++ json_serializable/test/annotation_version_test.dart
@@ -49,85 +49,66 @@ environment:
       );
     });
 
-    test(
-      'is at least the required `$supportedLanguageConstraint`',
-      () async => await _structurePackage(
+    test('is at least the required `$supportedLanguageConstraint`', () async {
+      await _structurePackage(
         dependencies: {'json_annotation': _annotationLowerBound},
         message: null,
-      ),
-    );
+      );
+    });
   });
 
-  test(
-    'missing dependency in production code',
-    () => _structurePackage(message: _missingProductionDep),
-  );
+  test('missing dependency in production code', () async {
+    await _structurePackage(message: _missingProductionDep);
+  });
 
-  test(
-    'missing dependency in example code',
-    () => _structurePackage(
+  test('missing dependency in example code', () async {
+    await _structurePackage(
       sourceDirectory: 'example',
       message:
           'You are missing a required dependency on json_annotation with a '
           'lower bound of at least "$_annotationLowerBound".',
-    ),
-  );
+    );
+  });
 
-  test(
-    'dev dependency with a production usage',
-    () => _structurePackage(
+  test('dev dependency with a production usage', () async {
+    await _structurePackage(
       devDependencies: {'json_annotation': _annotationLowerBound},
       message: _missingProductionDep,
-    ),
-  );
+    );
+  });
 
-  test(
-    'dependency with `null` constraint',
-    () => _structurePackage(
+  test('dependency with `null` constraint', () async {
+    await _structurePackage(
       dependencies: {'json_annotation': null},
       message:
           'The version constraint "any" on json_annotation allows versions '
           'before $_annotationLowerBound which is not allowed.',
-    ),
-  );
+    );
+  });
 
-  test(
-    'dependency with "any" constraint',
-    () => _structurePackage(
+  test('dependency with "any" constraint', () async {
+    await _structurePackage(
       dependencies: {'json_annotation': 'any'},
       message:
           'The version constraint "any" on json_annotation allows versions '
           'before $_annotationLowerBound which is not allowed.',
-    ),
-  );
+    );
+  });
 
-  test(
-    'dependency with too low version range',
-    () => _structurePackage(
+  test('dependency with too low version range', () async {
+    await _structurePackage(
       dependencies: {'json_annotation': '^4.0.0'},
       message:
           'The version constraint "^4.0.0" on json_annotation allows versions '
           'before $_annotationLowerBound which is not allowed.',
-    ),
-  );
+    );
+  });
 }
 
 final _jsonSerialPubspec = Pubspec.parse(
   File('pubspec.yaml').readAsStringSync(),
   sourceUrl: Uri.file('pubspec.yaml'),
 );
-
-String _fixPath(String path) {
-  if (p.isAbsolute(path)) return path;
-
-  return p.canonicalize(p.join(p.current, path));
-}
-
-final _jsonSerialPathDependencyOverrides = {
-  for (var entry in _jsonSerialPubspec.dependencyOverrides.entries)
-    if (entry.value is PathDependency)
-      entry.key: {'path': _fixPath((entry.value as PathDependency).path)},
-};
 
 final _annotationLowerBound = requiredJsonAnnotationMinVersion.toString();
 
@@ -154,7 +135,11 @@ Future<void> _structurePackage({
       'build_runner': 'any',
       'json_serializable': {'path': p.current},
     },
-    'dependency_overrides': _jsonSerialPathDependencyOverrides,
+    'dependency_overrides': {
+      'json_annotation': {
+        'path': p.canonicalize(p.join(p.current, '../json_annotation')),
+      },
+    },
   });
 
   await d.file('pubspec.yaml', pubspec).create();