+++ json_serializable/test/shared_config.dart
@@ -28,5 +28,6 @@ final generatorConfigNonDefaultJson = Map<String, dynamic>.unmodifiable(
     ignoreUnannotated: true,
     includeIfNull: false,
     genericArgumentFactories: true,
+    createJsonSchema: true,
   ).toJson(),
 );