+++ json_serializable/test/test_sources/test_sources.dart
@@ -24,6 +24,7 @@ class ConfigurationImplicitDefaults {
   ignoreUnannotated: false,
   includeIfNull: true,
   genericArgumentFactories: false,
+  createJsonSchema: false,
 )
 class ConfigurationExplicitDefaults {
   int? field;